Moved the currently in-use hellmouth levels to the volcano/ folder. Fixed the hellmouth1 level to be playable (still needs some tweaking on the release rate and time possibly) Volcano worldmap now loads and first level is playable (needs a lot of tweaking still, though) Switch/Door traps work again. Couple of other minor cleanups.

+++ src/worldmap/worldmap.cxx
@@ -216,7 +216,7 @@ WorldMap::draw (DrawingContext& gc)
                           _("...walking..."));
     }
 
-  Vector mpos = display_gc->screen_to_world(Vector(mouse_x, mouse_y));
+  Vector mpos = display_gc->screen_to_world(Vector((float)mouse_x, (float)mouse_y));
   Dot* dot = path_graph->get_dot(mpos.x, mpos.y);
   if (dot)
     {
